requirements.txt
babel==2.17.0
cchardet==2.2.0a2
coverage==4.3.1
coveralls==1.1
cython==3.0.12
datapackage==0.8.1
dj-database-url==0.5.0
django_bootstrap3==12.1.0
django-ckeditor==5.4.0
django-confy==1.0.4
django-countries==7.6.1
django-crispy-forms==2.0
django-cron==0.6.0 # This project is no longer maintained
django-datatables-view==1.20.0
django-dirtyfields==1.3.1
django-dynamic-fixture==1.9.1
django-extensions==3.2.3
django-extra-fields==0.9
django-phonenumber-field==8.0.0
django-reset-migrations==0.3.1
django-reversion-compare==0.15
django-reversion==5.1.0
django-widget-tweaks==1.5.0
django==5.0.13
djangorestframework-csv==3.0.2
djangorestframework-datatables==0.7.2
djangorestframework-gis==1.1
djangorestframework==3.15.2
geojson==1.3.3
git+https://github.com/dbca-wa/appmonitor_client.git#egg=appmonitor_client
git+https://github.com/dbca-wa/django5-utils-six.git#egg=django5_six
git+https://github.com/dbca-wa/ledger_api_client.git#egg=ledger_api_client
git+https://github.com/dbca-wa/webtemplate2.git#egg=webtemplate_dbca
gunicorn==19.9.0
ipython==8.32.0
jsontableschema==0.10.1
openpyxl==2.4.8
phonenumbers==8.12.15
psycopg2-binary==2.9.10
pycountry==17.1.8
PyPDF2==1.26.0
python-dateutil==2.9.0.post0
python-decouple==3.8
reportlab==4.3.1
unicodecsv==0.14.1
whitenoise==6.5.0
xlsxwriter==1.1.2
